refactor(GeometryCache): Moved from App to Common
