--- WolvenKit.App/PhysX/PhysXMesh.cs
@@ -1,6 +1,0 @@
-﻿namespace WolvenKit.App.PhysX;
-
-public class PhysXMesh
-{
-    
-}
