+++ WolvenKit.App/ViewModels/Documents/RDTMeshViewModel.cs
@@ -23,10 +23,10 @@ using WolvenKit.App.Controllers;
 using WolvenKit.App.Extensions;
 using WolvenKit.App.Helpers;
 using WolvenKit.App.Models;
-using WolvenKit.App.PhysX;
 using WolvenKit.App.Services;
 using WolvenKit.App.ViewModels.Shell;
 using WolvenKit.Common.Interfaces;
+using WolvenKit.Common.PhysX;
 using WolvenKit.Common.Services;
 using WolvenKit.Core.Extensions;
 using WolvenKit.Core.Interfaces;
