+++ WolvenKit.Common/PhysX/BV4Tree.cs
@@ -4,7 +4,7 @@ using System.IO;
 using SharpDX;
 using WolvenKit.Core.Extensions;
 
-namespace WolvenKit.App.PhysX;
+namespace WolvenKit.Common.PhysX;
 
 public class BV4Tree
 {
