+++ WolvenKit.Common/PhysX/ConvexMesh.cs
@@ -3,7 +3,7 @@ using System.IO;
 using SharpDX;
 using WolvenKit.Core.Extensions;
 
-namespace WolvenKit.App.PhysX;
+namespace WolvenKit.Common.PhysX;
 
 public class ConvexMesh : PhysXMesh
 {
