+++ WolvenKit.Common/PhysX/PhysXMesh.cs
@@ -1,0 +1,6 @@
+﻿namespace WolvenKit.Common.PhysX;
+
+public class PhysXMesh
+{
+    
+}
