+++ WolvenKit.Common/WolvenKit.Common.csproj
@@ -83,6 +83,7 @@
   <ItemGroup>
     <PackageReference Include="DynamicData" />
     <PackageReference Include="SharpDX.Direct3D11" />
+    <PackageReference Include="SharpDX.Mathematics" />
     <PackageReference Include="SharpGLTF.Core" />
     <PackageReference Include="SharpZipLib" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
